+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.0
+BREAKING CHANGES:
+
+Added onError option in CommonAsyncSnapshotResponses
+Updated onLoading option in CommonAsyncSnapshotResponses 
+Added onLoading option in FutureDialog.
+Standardized parameter names in FutureDialog, CommonAsyncSnapshotResponses to, onData, onLoading
+and onError.
+
 ## 0.0.3
 
 Added debug and throwError option for FutureDialog and CommonAsyncSnapshotResponses.